+++ Changelog.md
@@ -8,6 +8,10 @@
   - Adds a string list constructor.
   - Add an interface to determine if connected: `isConnected()`.
   - Add an interface to determine if data synchronization has started: `isStarted()`.
+- `DashboardClient`
+  - Added interface to get robot type: `robotType()`
+  - Added interface to get robot serial number: `robotSerialNumber()`
+  - Added interface to get robot ID: `robotID()`
 
 ### Changed
 - Adjust the "timeout" value of "trajectory_socket" in `external_control.script`.
@@ -19,6 +23,9 @@
 - Fix the issue where the error level was incorrect when parsing the 30001 robot error exception message.
 - Fix the memory issue that occurs after receiving incomplete messages from 30001.
 - Fix the issue where static resources in `TcpServer` are destructed prematurely.
+
+### Deprecated
+- Deprecated `DashboardClient::robot()` it will be removed in future versions. Please use `DashboardClient::robotType()` instead.
 
 ## [v1.2.0] - 2025-08-14